move element selectors to helper func file
