+++ src/components/IsSolutionCorrect.js
@@ -1,3 +1,4 @@
+import {get_row, get_col, get_square} from '../common/helper_functions';
 // 'checkGrid should be called - outputs a boolean confirmation if the sudoku grid meets completion criteria.
 
 class IsSolutionCorrect {
@@ -7,7 +8,7 @@ class IsSolutionCorrect {
     
     checkRows(grid) {
         for (let rowIndex = 0; rowIndex < 9; rowIndex++) {
-            let row = grid[rowIndex];
+            let row = get_row(grid, rowIndex);
             if (!this.containsOneToNine(row)) { return false }
         }
         return true;
@@ -15,7 +16,7 @@ class IsSolutionCorrect {
     
     checkColumns(grid) {
         for (let columnIndex = 0; columnIndex < 9; columnIndex++) {
-            let column = grid.map(row => row[columnIndex]);
+            let column = get_col(grid, columnIndex);
             if (!this.containsOneToNine(column)) { return false }
         }
         return true;
@@ -23,9 +24,7 @@ class IsSolutionCorrect {
     
     checkSquares(grid) {
         for (let squareIndex = 0; squareIndex < 9; squareIndex++) {
-            let square = grid.filter((val, rowIndex) => Math.floor(rowIndex / 3) === Math.floor(squareIndex / 3))
-                             .map(x => x.slice((squareIndex % 3) * 3, (squareIndex % 3) * 3 + 3))
-                             .flat();
+            let square = get_square(grid, squareIndex)
             if (!this.containsOneToNine(square)) { return false }
         }
         return true;
